+++ dawn/test/unit-test/dawn/CodeGen/Naive/TestCodeGenNaive.cpp
@@ -19,7 +19,7 @@
 #include "dawn/SIR/SIR.h"
 #include "dawn/Support/DiagnosticsEngine.h"
 #include "dawn/Support/FileUtil.h"
-#include "dawn/Unittest/CodeDumper.h"
+#include "dawn/Unittest/CompilerUtil.h"
 #include "dawn/Unittest/IIRBuilder.h"
 #include "dawn/Unittest/UnittestLogger.h"
 
@@ -48,7 +48,7 @@ TEST(CodeGenNaiveTest, GlobalIndexStencil) {
                                      b.block(b.stmt(b.assignExpr(b.at(out_f), b.lit(10)))))))));
 
   std::ostringstream oss;
-  dawn::CodeDumper::dumpNaive(oss, stencil_instantiation);
+  dawn::CompilerUtil::dumpNaive(oss, stencil_instantiation);
   std::string gen = oss.str();
 
   std::string ref = dawn::readFile("reference/global_indexing.cpp");
@@ -87,7 +87,7 @@ TEST(CodeGenNaiveTest, NonOverlappingInterval) {
                              b.block(b.stmt(b.assignExpr(b.at(out), b.lit(10)))))))));
 
   std::ostringstream oss;
-  dawn::CodeDumper().dumpNaive(oss, stencil_inst);
+  dawn::CompilerUtil().dumpNaive(oss, stencil_inst);
   std::string gen = oss.str();
 
   std::string ref = dawn::readFile("reference/nonoverlapping_stencil.cpp");
@@ -124,7 +124,7 @@ TEST(CodeGenNaiveTest, LaplacianStencil) {
                       b.binaryExpr(b.at(dx), b.at(dx), Op::multiply), Op::divide)))))))));
 
   std::ofstream ofs("test/unit-test/dawn/CodeGen/Naive/generated/laplacian_stencil.cpp");
-  dawn::CodeDumper::dumpNaive(ofs, stencil_inst);
+  dawn::CompilerUtil::dumpNaive(ofs, stencil_inst);
 }
 
 } // anonymous namespace
