+++ dawn/test/unit-test/dawn/Validator/CMakeLists.txt
@@ -20,7 +20,7 @@ add_executable(${executable}
   TestGridTypeChecker.cpp
   TestIntegrityChecker.cpp
 )
-target_link_libraries(${executable} DawnCompiler DawnValidator DawnUnittest gtest)
+target_link_libraries(${executable} DawnValidator DawnUnittest gtest)
 target_add_dawn_standard_props(${executable})
 gtest_discover_tests(${executable} TEST_PREFIX "Dawn::Unit::Validator::")
 file(COPY input DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
